#10 [chore] : directory 위치 변경
