+++ app/src/main/java/com/byeboo/app/presentation/offbording/component/OffboardingModal.kt
@@ -1,4 +1,4 @@
-package com.byeboo.app.core.designsystem.component.modal
+package com.byeboo.app.presentation.offbording.component
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
@@ -6,9 +6,9 @@ import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Text
@@ -17,11 +17,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.SpanStyle
-import androidx.compose.ui.text.buildAnnotatedString
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.text.style.TextDecoration
-import androidx.compose.ui.text.withStyle
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.byeboo.app.R
@@ -29,10 +25,10 @@ import com.byeboo.app.core.designsystem.component.button.ByeBooButton
 import com.byeboo.app.core.designsystem.ui.theme.ByeBooTheme
 
 @Composable
-fun QuestModal(
-    questNumber: String,
-    questQuestion: String,
+fun OffboardingModal(
     onClick: () -> Unit,
+    titleText: String,
+    guideText: String,
     modifier: Modifier = Modifier
 ) {
     Box(
@@ -40,25 +36,15 @@ fun QuestModal(
             .width(264.dp)
             .clip(RoundedCornerShape(12.dp))
             .background(color = ByeBooTheme.colors.gray900Alpha80)
-            .padding(24.dp),
+            .padding(24.dp)
     ) {
         Column(
             modifier = Modifier,
             horizontalAlignment = Alignment.CenterHorizontally,
             verticalArrangement = Arrangement.Center
         ) {
-            Image(
-                painter = painterResource(id = R.drawable.bori_quest_banner),
-                contentDescription = "이미지",
-                Modifier
-                    .fillMaxWidth()
-                    .height(58.dp)
-            )
-
-            Spacer(modifier = Modifier.height((17.5).dp))
-
             Text(
-                text = questNumber,
+                text = "축하드려요!",
                 style = ByeBooTheme.typography.body2,
                 color = ByeBooTheme.colors.gray400
             )
@@ -66,7 +52,7 @@ fun QuestModal(
             Spacer(modifier = Modifier.height(8.dp))
 
             Text(
-                text = questQuestion,
+                text = titleText,
                 style = ByeBooTheme.typography.sub2,
                 color = ByeBooTheme.colors.gray50,
                 textAlign = TextAlign.Center
@@ -74,21 +60,26 @@ fun QuestModal(
 
             Spacer(modifier = Modifier.height(16.dp))
 
-            Text(
-                text = buildAnnotatedString {
-                    withStyle(style = SpanStyle(textDecoration = TextDecoration.Underline)) {
-                        append(text = "작성 TIP")
-                    }
-                },
-                style = ByeBooTheme.typography.body4,
-                color = ByeBooTheme.colors.gray300
+            Image(
+                painter = painterResource(id = R.drawable.bori_clover),
+                contentDescription = "이미지",
+                modifier = Modifier.size(160.dp)
             )
 
-            Spacer(modifier = Modifier.height((17.5).dp))
+            Spacer(modifier = Modifier.height(16.dp))
+
+            Text(
+                text = guideText,
+                style = ByeBooTheme.typography.body2,
+                color = ByeBooTheme.colors.gray400,
+                textAlign = TextAlign.Center
+            )
+
+            Spacer(modifier = Modifier.height(16.dp))
 
             ByeBooButton(
                 onClick = onClick,
-                buttonText = "진행하기",
+                buttonText = "바로가기",
                 buttonTextColor = ByeBooTheme.colors.white,
                 buttonBackgroundColor = ByeBooTheme.colors.primary300,
             )
@@ -98,12 +89,12 @@ fun QuestModal(
 
 @Preview(showBackground = true)
 @Composable
-fun QuestModalPreview() {
+fun OffboardingModalPreview() {
     ByeBooTheme {
-        QuestModal(
-            questNumber = "열 번째 퀘스트",
-            questQuestion = "연애에서 반복됐던 문제 패턴\n3가지를 생각해보아요.",
-            onClick = {}
+        OffboardingModal(
+            onClick = {},
+            titleText = "감정 직면 여정을\n모두 마무리 했어요",
+            guideText = "보리가 하츠핑님께\n하고싶은 말이 있다고 해요"
         )
     }
 }
