+++ app/src/main/java/com/byeboo/app/presentation/quest/component/QuitModal.kt
@@ -1,4 +1,4 @@
-package com.byeboo.app.core.designsystem.component.modal
+package com.byeboo.app.presentation.quest.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
